Clean up implicit downcasts
Clean up implicit downcasts by:
* Adding explicit casts.
* Changing `GenericClass` to `GenericClass<T>` in a few more places
  to avoid implicitly using `dynamic`.
* Removing explicit `GenericClass` declarations in other places so
  that the type is inferred instead of being retyped to
  `GenericClass<dynamic>`.
* Changing `Derived object = baseObject;` to
  `var object = baseObject as Derived`.
* Adding lots of `.toDouble()` calls to fix implicit casts from
  `num`.  Some of these sites maybe could use `as double` instead,
  but using `.toDouble()` is safer and easier to reason about.

PiperOrigin-RevId: 363460124

+++ charts_common/lib/src/chart/bar/bar_error_decorator.dart
@@ -80,10 +80,11 @@ class BarErrorDecorator<D> extends BarRendererDecorator<D> {
           if (startY != endY) {
             final barWidth = bounds.right - bounds.left;
             final x = (bounds.left + bounds.right) / 2;
-            final rectWidth = min(strokeWidthPx + 2 * outlineWidthPx, barWidth);
+            final rectWidth =
+                min(strokeWidthPx + 2 * outlineWidthPx, barWidth.toDouble());
             final strokeWidth = rectWidth - 2 * outlineWidthPx;
             final rectEndpointLength =
-                min(endpointLengthPx + 2 * outlineWidthPx, barWidth);
+                min(endpointLengthPx + 2 * outlineWidthPx, barWidth.toDouble());
             final endpointLength = rectEndpointLength - 2 * outlineWidthPx;
 
             if (outlineWidthPx > 0) {
@@ -138,10 +139,11 @@ class BarErrorDecorator<D> extends BarRendererDecorator<D> {
           if (startX != endX) {
             final barWidth = bounds.bottom - bounds.top;
             final y = (bounds.top + bounds.bottom) / 2;
-            final rectWidth = min(strokeWidthPx + 2 * outlineWidthPx, barWidth);
+            final rectWidth =
+                min(strokeWidthPx + 2 * outlineWidthPx, barWidth.toDouble());
             final strokeWidth = rectWidth - 2 * outlineWidthPx;
             final rectEndpointLength =
-                min(endpointLengthPx + 2 * outlineWidthPx, barWidth);
+                min(endpointLengthPx + 2 * outlineWidthPx, barWidth.toDouble());
             final endpointLength = rectEndpointLength - 2 * outlineWidthPx;
 
             if (outlineWidthPx > 0) {
