+++ charts_common/lib/src/chart/cartesian/axis/linear/bucketing_numeric_axis.dart
@@ -63,13 +63,15 @@ class BucketingNumericAxis extends NumericAxis {
   /// Gets the location of [domain] on the axis, repositioning any value less
   /// than [threshold] to the middle of the reserved bucket.
   @override
-  double getLocation(num domain) {
+  double /*?*/ getLocation(num domain) {
     if (domain == null) {
       return null;
     } else if (_threshold != null && domain < _threshold) {
-      return _showBucket ? scale[_threshold / 2] : scale[0.0];
+      return (_showBucket
+          ? scale[_threshold / 2]
+          : scale[0.0]) /*!*/ .toDouble();
     } else {
-      return scale[domain];
+      return scale[domain] /*!*/ .toDouble();
     }
   }
 }
