+++ charts_common/lib/src/chart/cartesian/axis/numeric_extents.dart
@@ -80,10 +80,10 @@ class NumericExtents implements Extents<num> {
 
   // Returns true if these [NumericExtents] collides with [other].
   bool overlaps(NumericExtents other) {
-    return _containsValue(other.min) ||
-        _containsValue(other.max) ||
-        other._containsValue(min) ||
-        other._containsValue(max);
+    return _containsValue(other.min.toDouble()) ||
+        _containsValue(other.max.toDouble()) ||
+        other._containsValue(min.toDouble()) ||
+        other._containsValue(max.toDouble());
   }
 
   @override
